Renames action count to turn count
Renames the function to more accurately reflect its purpose, which is to count main turns rather than just actions.

Adds documentation to explain the concept of "main turns" and how they are calculated.
